Move datadog.trace.api.utils to datadog.trace.util
