+++ internal-api/src/test/groovy/datadog/trace/util/MathUtilsTest.groovy
@@ -1,4 +1,5 @@
-package datadog.trace.api.utils
+package datadog.trace.util
+
 
 import datadog.trace.test.util.DDSpecification
